Combine GraphicalUserInterface and Ui into one
