+++ src/main/java/duke/Ui.java
@@ -1,11 +1,26 @@
 package duke;
 
 
-import java.util.Scanner;
-
 import duke.task.Task;
 import duke.task.TaskList;
 import javafx.application.Platform;
+import javafx.geometry.Insets;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.Background;
+import javafx.scene.layout.BackgroundFill;
+import javafx.scene.layout.Region;
+import javafx.scene.layout.VBox;
+import javafx.scene.paint.Paint;
+import javafx.scene.shape.Circle;
+import javafx.stage.Stage;
+
 
 /**
  * Class responsible for all UI interactions with user.
@@ -14,27 +29,62 @@ import javafx.application.Platform;
  */
 public class Ui {
 
-    private Scanner scanner;
-    private boolean isGui = false;
-    private GraphicalUserInterface gui = null;
+    private Stage stage;
+    private ScrollPane scrollPane;
+    private VBox dialogContainer;
+    private AnchorPane mainLayout;
+    private TextField userInput;
+    private Button sendButton;
+    private Scene scene;
+    private Image user = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
+    private Image duke = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
+
+
+    /**
+     * Constructor for jUnit testing
+     */
+    public Ui() {
+
+    }
 
     /**
      * Constructor for Ui class
      */
-    public Ui() {
-        scanner = new Scanner(System.in);
+    public Ui(Stage stage) {
+        this.stage = stage;
+
+        // setting variables
+        scrollPane = new ScrollPane();
+        dialogContainer = new VBox();
+        scrollPane.setContent(dialogContainer);
+        userInput = new TextField();
+        sendButton = new Button("Send");
+        mainLayout = new AnchorPane();
+        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
+        scene = new Scene(mainLayout);
+
+        // setup styling/ formatting
+        stage.setScene(scene);
+        stage.setTitle("Duke");
+        stage.setResizable(false);
+        stage.setMinHeight(600.0);
+        stage.setMinWidth(400.0);
+        mainLayout.setPrefSize(400.0, 600.0);
+        scrollPane.setPrefSize(385, 535);
+        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
+        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
+        scrollPane.setVvalue(1.0);
+        scrollPane.setFitToWidth(true);
+        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
+        userInput.setPrefWidth(325.0);
+        sendButton.setPrefWidth(55.0);
+        AnchorPane.setTopAnchor(scrollPane, 1.0);
+        AnchorPane.setBottomAnchor(sendButton, 1.0);
+        AnchorPane.setRightAnchor(sendButton, 1.0);
+        AnchorPane.setLeftAnchor(userInput, 1.0);
+        AnchorPane.setBottomAnchor(userInput, 1.0);
     }
 
-    /**
-     * Allows a GraphicalUserInterface to be added to the Ui so that appropriate GUI responses can be generated.
-     *
-     * @param gui The GUI to be added.
-     */
-    public void addGui(GraphicalUserInterface gui) {
-        scanner.close();
-        this.gui = gui;
-        this.isGui = true;
-    }
 
     /**
      * Displays the message of any DukeException thrown by Duke.
@@ -43,11 +93,8 @@ public class Ui {
      */
     protected void showException(DukeException e) {
         String message = e.getMessage() + "\n";
-        if (isGui) {
-            gui.displayDukeReply(message);
-            return;
-        }
-        displayMessage(message);
+        displayDukeReply(message);
+
     }
 
     /**
@@ -55,11 +102,8 @@ public class Ui {
      */
     protected void greet() {
         String message = "Good Day Sir/Mdm, I am Duke.\nWhat can I do for you?\n";
-        if (isGui) {
-            gui.displayDukeReply(message);
-            return;
-        }
-        displayMessage(message);
+        displayDukeReply(message);
+
     }
 
     /**
@@ -67,13 +111,9 @@ public class Ui {
      */
     public void end() {
         String message = "Goodbye Sir/Mdm. Hope to serve you again soon!\n";
-        if (isGui) {
-            gui.displayDukeReply(message);
-            Platform.exit();
-            return;
-        }
-        scanner.close();
-        displayMessage(message);
+        displayDukeReply(message);
+        Platform.exit();
+
     }
 
     /**
@@ -85,11 +125,8 @@ public class Ui {
     public void add(Task task, TaskList tasks) {
         String message = "Understood Sir/Mdm, I have added the indicated task: " + "\n   " + task + "\n"
             + "Now you have " + tasks.size() + (tasks.size() == 1 ? " task." : " tasks.") + "\n";
-        if (isGui) {
-            gui.displayDukeReply(message);
-            return;
-        }
-        displayMessage(message);
+        displayDukeReply(message);
+
     }
 
     /**
@@ -99,11 +136,7 @@ public class Ui {
      */
     public void showList(TaskList tasks) {
         String message = "Here are your tasks Sir/Mdm:\n" + list(tasks) + "\n";
-        if (isGui) {
-            gui.displayDukeReply(message);
-            return;
-        }
-        displayMessage(message);
+        displayDukeReply(message);
     }
 
     /**
@@ -114,11 +147,8 @@ public class Ui {
     public void markDone(Task task) {
         String message = "Good job Sir/Mdm! I shall mark this task as complete:\n   "
             + task + "\n";
-        if (isGui) {
-            gui.displayDukeReply(message);
-            return;
-        }
-        displayMessage(message);
+        displayDukeReply(message);
+
     }
 
     /**
@@ -132,11 +162,10 @@ public class Ui {
             + task + "\n" + "Now you have " + tasks.size()
             + (tasks.size() == 1 ? " task." : " tasks.") + "\n";
 
-        if (isGui) {
-            gui.displayDukeReply(message);
-            return;
-        }
-        displayMessage(message);
+
+        displayDukeReply(message);
+
+
     }
 
     /**
@@ -161,11 +190,9 @@ public class Ui {
     public void findByDate(TaskList foundTasks) {
         String message = "Here are the deadlines and events that match the date Sir/Mdm:\n"
             + list(foundTasks) + "\n";
-        if (isGui) {
-            gui.displayDukeReply(message);
-            return;
-        }
-        displayMessage(message);
+
+        displayDukeReply(message);
+
     }
 
     /**
@@ -176,29 +203,79 @@ public class Ui {
     public void findByDescription(TaskList foundTasks) {
         String message = "Here are the results of the search Sir/Mdm:\n"
             + list(foundTasks) + "\n";
-        if (isGui) {
-            gui.displayDukeReply(message);
-            return;
-        }
-        displayMessage(message);
+
+        displayDukeReply(message);
     }
 
     /**
-     * Reads the user's input from the Command Line Interface.
-     *
-     * @return The user's input.
-     * @throws DukeException
+     * Display GUI and greet user.
      */
-    protected String readCommand() throws DukeException {
-        if (isGui) {
-            throw new DukeException("Something wrong happened: Cannot read command from GUI");
+    public void showGui(TaskList tasks, Storage storage) {
+        // Create event listeners
+        sendButton.setOnMouseClicked((event) -> {
+            handleUserInput(tasks, storage);
+        });
+
+        userInput.setOnAction((event) -> {
+            handleUserInput(tasks, storage);
+        });
+        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
+
+        stage.show();
+        this.greet();
+    }
+
+
+    private void handleUserInput(TaskList tasks, Storage storage) {
+
+        String userInputText = userInput.getText();
+        Label userTextLabel = new Label(userInputText);
+        userTextLabel.setPadding(new Insets(0, 10, 0, 0));
+        ImageView userImage = new ImageView(user);
+        clipImageViewToCircle(userImage, 100);
+        DialogBox userDialogBox = DialogBox.getUserDialog(userTextLabel, userImage);
+        userDialogBox.setPadding(new Insets(0, 0, 30, 0));
+        setDialogBoxBackgroundColor(userDialogBox, "#1FDA12");
+
+        dialogContainer.getChildren().add(
+            userDialogBox
+        );
+
+        try {
+            Parser.parse(userInputText, this, tasks, storage).execute();
+        } catch (DukeException e) {
+            this.showException(e);
         }
-        return scanner.nextLine();
-    }
 
 
-    private static void displayMessage(String message) {
-        System.out.println(message);
+        userInput.clear();
     }
+
+    private void displayDukeReply(String dukeReply) {
+        Label dukeTextLabel = new Label(dukeReply);
+        dukeTextLabel.setPadding(new Insets(0, 0, 0, 10));
+
+        ImageView dukeImage = new ImageView(duke);
+        clipImageViewToCircle(dukeImage, 100);
+
+        DialogBox dukeDialogBox = DialogBox.getDukeDialog(dukeTextLabel, dukeImage);
+        dukeDialogBox.setPadding(new Insets(0, 0, 30, 0));
+
+        setDialogBoxBackgroundColor(dukeDialogBox, "#12B1DA");
+
+        dialogContainer.getChildren().add(dukeDialogBox);
+    }
+
+    private void clipImageViewToCircle(ImageView imageView, int length) {
+        imageView.setPreserveRatio(false);
+        imageView.setSmooth(true);
+        Circle circle = new Circle(length / 2, length / 2, Math.min(length, length) / 2);
+        imageView.setClip(circle);
+    }
+
+    private void setDialogBoxBackgroundColor(DialogBox dialogBox, String hexValue) {
+        dialogBox.setBackground(new Background(new BackgroundFill(Paint.valueOf(hexValue), null, null)));
+    }
+
 
 }
